U-106 - Separated time service into 3 services (time, day, month)
