+++ src/app/modules/course/modals/period-modal/period-modal.component.ts
@@ -3,8 +3,8 @@ import {FormBuilder, FormControl, FormGroup, Validators} from '@angular/forms';
 import {NgbActiveModal} from '@ng-bootstrap/ng-bootstrap';
 import {ButtonText} from '../../../../shared/enums/button-text.enum';
 import {Period} from '../../../../models/period.model';
-import {TimeService} from '../../../../core/services/time.service';
 import {ComboBoxItem} from '../../../../shared/models/combo-box.model';
+import {MonthService} from '../../../../core/services/time/month.service';
 
 @Component({
   selector: 'app-period-modal',
@@ -31,7 +31,7 @@ export class PeriodModalComponent implements OnInit {
   }
 
   private getListOfMonths() {
-    this.months = TimeService.getListOfMonths();
+    this.months = MonthService.getListOfMonths();
   }
 
   public cancelAction(): void {
